ConDec-428: Unify test classes for createLink and deleteLink methods in KnowledgeRest
