--- src/test/java/de/uhd/ifi/se/decision/management/jira/rest/knowledgerest/TestCreateGenericLink.java
@@ -1,75 +1,0 @@
-package de.uhd.ifi.se.decision.management.jira.rest.knowledgerest;
-
-import static org.junit.Assert.assertEquals;
-
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import com.google.common.collect.ImmutableMap;
-
-import de.uhd.ifi.se.decision.management.jira.TestSetUpWithIssues;
-import de.uhd.ifi.se.decision.management.jira.extraction.model.Comment;
-import de.uhd.ifi.se.decision.management.jira.extraction.model.TestComment;
-import de.uhd.ifi.se.decision.management.jira.extraction.persistence.ActiveObjectsManager;
-import de.uhd.ifi.se.decision.management.jira.model.Link;
-import de.uhd.ifi.se.decision.management.jira.persistence.GenericLinkManager;
-import net.java.ao.test.jdbc.Data;
-import net.java.ao.test.jdbc.NonTransactional;
-import net.java.ao.test.junit.ActiveObjectsJUnitRunner;
-
-@RunWith(ActiveObjectsJUnitRunner.class)
-@Data(TestSetUpWithIssues.AoSentenceTestDatabaseUpdater.class)
-public class TestCreateGenericLink extends TestKnowledgeRestSetUp {
-
-	private final static String CREATION_ERROR = "Creation of link failed.";
-
-	private Link newGenericLink() {
-		TestComment tc = new TestComment();
-		Comment comment = tc.getComment("first Comment");
-		long id = ActiveObjectsManager.addNewSentenceintoAo(comment, comment.getIssueId(), 0);
-		
-		return GenericLinkManager.getLinksForElement("s"+id).get(0);
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestNullElementNull() {
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.createLink(null, null, null).getEntity());
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestNullElementFilled() {
-		TestComment tc = new TestComment();
-		Comment comment = tc.getComment("this is atest sentence");
-		decisionKnowledgeElement = comment.getSentences().get(0);
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.createLink("TEST", null, newGenericLink()).getEntity());
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestFilledElementNull() {
-		request.setAttribute("WithFails", false);
-		request.setAttribute("NoFails", true);
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.createLink("TEST", request, null).getEntity());
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestFilledElementFilled() {
-		request.setAttribute("WithFails", false);
-		request.setAttribute("NoFails", true);
-		Link gl = newGenericLink();
-		gl.setId(0);
-		GenericLinkManager.insertLink(gl, null);
-		assertEquals(Status.OK.getStatusCode(),
-				knowledgeRest.createLink("TEST", request, newGenericLink()).getStatus());
-	}
-
-}
