--- src/test/java/de/uhd/ifi/se/decision/management/jira/rest/knowledgerest/TestDeleteGenericLink.java
@@ -1,119 +1,0 @@
-package de.uhd.ifi.se.decision.management.jira.rest.knowledgerest;
-
-import static org.junit.Assert.assertEquals;
-
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import com.google.common.collect.ImmutableMap;
-
-import de.uhd.ifi.se.decision.management.jira.TestSetUpWithIssues;
-import de.uhd.ifi.se.decision.management.jira.extraction.model.Comment;
-import de.uhd.ifi.se.decision.management.jira.extraction.model.TestComment;
-import de.uhd.ifi.se.decision.management.jira.model.Link;
-import de.uhd.ifi.se.decision.management.jira.model.LinkImpl;
-import de.uhd.ifi.se.decision.management.jira.persistence.GenericLinkManager;
-import net.java.ao.test.jdbc.Data;
-import net.java.ao.test.jdbc.NonTransactional;
-import net.java.ao.test.junit.ActiveObjectsJUnitRunner;
-
-@RunWith(ActiveObjectsJUnitRunner.class)
-@Data(TestSetUpWithIssues.AoSentenceTestDatabaseUpdater.class)
-public class TestDeleteGenericLink extends TestKnowledgeRestSetUp {
-
-	private final static String CREATION_ERROR = "Deletion of link failed.";
-
-	@Test
-	@NonTransactional
-	public void testRequestNullElementNull() {
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.deleteLink(null, null, null).getEntity());
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestNullElementFilled() {
-		assertEquals(
-				Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-						.getEntity(),
-				knowledgeRest.deleteLink("TEST", null, new LinkImpl(decisionKnowledgeElement, sentence))
-						.getEntity());
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestFilledElementNull() {
-		request.setAttribute("WithFails", false);
-		request.setAttribute("NoFails", true);
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.deleteLink("TEST", request, null).getEntity());
-	}
-
-	@Test
-	@NonTransactional
-	public void testRequestFilledElementFilled() {
-		request.setAttribute("WithFails", false);
-		request.setAttribute("NoFails", true);
-
-		TestComment tc = new TestComment();
-		Comment comment = tc.getComment("this is atest sentence");
-		sentence = comment.getSentences().get(0);
-
-		Link newLink = new LinkImpl(decisionKnowledgeElement, sentence);
-		GenericLinkManager.insertLink(newLink, null);
-		
-		//Test Element existing in AO
-		assertEquals(1, GenericLinkManager
-				.getLinksForElement("i" + decisionKnowledgeElement.getId()).size());
-
-		assertEquals(Status.OK.getStatusCode(), knowledgeRest.deleteLink("TEST", request, newLink).getStatus());
-
-		GenericLinkManager.insertLink(newLink, null);
-		assertEquals(Status.OK.getStatusCode(), knowledgeRest.deleteLink("TEST", request, newLink).getStatus());
-
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.deleteLink("TEST", request, newLink).getEntity());
-		
-		//Test element not more exisitng in AO
-		assertEquals(0, GenericLinkManager
-				.getLinksForElement("i" + decisionKnowledgeElement.getId()).size());
-
-	}
-	
-	@Test
-	@NonTransactional
-	public void testRequestFilledElementFilledWithFlippedLink() {
-		request.setAttribute("WithFails", false);
-		request.setAttribute("NoFails", true);
-
-		TestComment tc = new TestComment();
-		Comment comment = tc.getComment("this is atest sentence");
-		sentence = comment.getSentences().get(0);
-
-		Link newLink = new LinkImpl(decisionKnowledgeElement, sentence);
-		GenericLinkManager.insertLink(newLink, null);
-		//Flip link
-		newLink = new LinkImpl(sentence,decisionKnowledgeElement);
-		
-		//Test Element existing in AO
-		assertEquals(1, GenericLinkManager
-				.getLinksForElement("i" + decisionKnowledgeElement.getId()).size());
-
-		assertEquals(Status.OK.getStatusCode(), knowledgeRest.deleteLink("TEST", request, newLink).getStatus());
-
-		GenericLinkManager.insertLink(newLink, null);
-		assertEquals(Status.OK.getStatusCode(), knowledgeRest.deleteLink("TEST", request, newLink).getStatus());
-
-		assertEquals(Response.status(Status.BAD_REQUEST).entity(ImmutableMap.of("error", CREATION_ERROR)).build()
-				.getEntity(), knowledgeRest.deleteLink("TEST", request, newLink).getEntity());
-		
-		//Test element not more exisitng in AO
-		assertEquals(0, GenericLinkManager
-				.getLinksForElement("i" + decisionKnowledgeElement.getId()).size());
-
-	}
-
-}
